backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
pymysql==1.1.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
alibabacloud-dingtalk>=2.2.0
